requirements.txt
numpy>=1.24.0
openai>=1.12.0
python-dotenv>=1.0.0
anthropic>=0.18.0
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
streamlit>=1.30.0
plotly>=5.18.0
psycopg2-binary>=2.9.9
pandas>=2.1.0